+++ modules/services/views/src/main/java/au/com/shiftyjelly/pocketcasts/views/multiselect/MultiSelectHelper.kt
@@ -208,4 +208,10 @@ abstract class MultiSelectHelper<T> : CoroutineScope {
             }
         }
     }
+
+    fun cleanup() {
+        listener = null
+        coordinatorLayout = null
+        context = null
+    }
 }